pyproject.toml
[tool.poetry]
name = "wkseguros"
version = "0.1.0"
description = ""
authors = ["bywilliams <willian.seu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
pydantic = "^2.9.2"
passlib = "^1.7.4"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
email-validator = "^2.2.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.6.8"
httpx = "^0.27.2"

[tool.ruff] 
line-length = 79 
extend-exclude = ['migrations']

[tool.ruff.lint] 
preview = true 
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format] 
preview = true 
quote-style = 'single'

[tool.pytest.ini_options] 
pythonpath = "." 
addopts = '-p no:warnings'

[tool.taskipy.tasks] 
lint = 'ruff check .; ruff check . --diff' 
format = 'ruff check . --fix; ruff format .' 
run = "FLASK_APP=wkseguros/frontend/app.py flask run --debug"
pre_test = 'task lint' 
test = 'pytest -s -x --cov=wkseguros -vv' 
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

